requirements.txt
Django==3.2.23
Unipath==1.1
coverage==4.5.4
django-cors-headers==3.14.0
django-discover-runner==1.0
django-redis==4.12.1
django-cachalot==2.4.2
djangorestframework==3.14.0
gunicorn==21.2.0
mimeparse==0.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-mimeparse==1.6.0
drf-spectacular==0.27.2

python-dotenv~=0.19.0
requests~=2.26.0
pyuca~=1.2
icu~=0.0.1